refactor: extract shared clipboard write utility with ios safari fallback
Move the clipboard write-with-fallback logic into a reusable
writeToClipboard() function in clipboard.ts. Both copyMessageContent$
and copyToClipboard$ now use this shared utility, fixing the unhandled
rejection bug in copyToClipboard$ as well.
