+++ turbo/apps/platform/src/signals/zero-page/zero-session-chat-ui.ts
@@ -1,5 +1,5 @@
 import { state, computed, command } from "ccstate";
-import { throwIfAbort } from "../utils.ts";
+import { writeToClipboard } from "./clipboard.ts";
 
 // ---------------------------------------------------------------------------
 // Collapsible timeline expanded state
@@ -36,29 +36,11 @@ export const copiedMessageIdValue$ = computed((get) => {
 
 export const copyMessageContent$ = command(
   async ({ set }, messageId: string, content: string, signal: AbortSignal) => {
-    try {
-      await navigator.clipboard.writeText(content);
-    } catch (error: unknown) {
-      throwIfAbort(error);
-      // Clipboard API can throw NotAllowedError on iOS Safari when the user
-      // gesture context is lost (e.g. after an async boundary). Fall back to
-      // the legacy execCommand approach.
-      try {
-        const textarea = document.createElement("textarea");
-        textarea.value = content;
-        textarea.style.position = "fixed";
-        textarea.style.opacity = "0";
-        document.body.appendChild(textarea);
-        textarea.select();
-        document.execCommand("copy");
-        textarea.remove();
-      } catch (fallbackError: unknown) {
-        throwIfAbort(fallbackError);
-        // Both methods failed — nothing more we can do.
-        return;
-      }
-    }
+    const ok = await writeToClipboard(content);
     signal.throwIfAborted();
+    if (!ok) {
+      return;
+    }
     set(copiedMessageId$, messageId);
     window.setTimeout(() => {
       return set(copiedMessageId$, null);
